Handle not_my_vbucket errors in get_{tap,dcp}_docs_estimate.
This is needed after the subsequent change that change quick_stats not
to ignore errors.

ns_rebalance_observer doesn't bother to check if vbucket exists on a
node before getting the estimates. It's probably better to handle this
on the caller's side, but since these functions are exposed via
janitor_agent, that would require ensuring that master always runs
4.5.1, which I don't want to do at the moment.

Other kinds of errors are unexpected (or at least it seems so to me),
so it's fine to crash on those.

Change-Id: I00baf4614c08b335d250bc3e11bb98aff064a93b
Reviewed-on: http://review.couchbase.org/64345
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>