MB-20162 Ignore size threshold when compacting master db.
Fragmentation size threshold is a value that is supposed to be mean
the max absolute wasted space per bucket per node, so it doesn't make
sense on a per vbucket level (unless scaled properly).

And given that master vbuckets are quite special, we actually cheat and
ignore it altogether.

Change-Id: I36cdbace91903456a9f68e64788f484b871af59d
Reviewed-on: http://review.couchbase.org/65784
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>