MB-20595: Update cbdocloader call to support new docloader
The new cbdocloader is backwards compatibile with the old one
except we do not accept the BUCKET_USERNAME and BUCKET_PASSWORD
environment variables because they were not needed even in the
old cbdocloader.

We also added a flag for verbose output and to use two client
threads when loading sample data.

Change-Id: I5665090d3562e3c4d654c6631631600914b1c7fe
Reviewed-on: http://review.couchbase.org/66900
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>