MB-20682: Bad arity fix in the computed stat extractor routine.
A user request for "show by server" stats page for DCP QUEUES |
VIEWS/INDEXES would result in a badarity exception as the Views and
Indexes stats tries to combine 3 different values. But the combiner
callback only accepts 2 and hence the exception.

The patch is mostly proposed by Poonam. The following are the
changes to fix the bug:
[1] The combiner callback will now to accept 3 parameters.
[2] The combine_samples function will now accumulate all the
    stats and then call the combiner callback with all the needed
    parameters.

Change-Id: Ie7905227e5b499defeea8e303a983276dab81eed
Reviewed-on: http://review.couchbase.org/67191
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>