spock-ui: MB-20824: remove unnecessary <base> tag
since we no longer need to override a default URL
and a default target for all links on a page. Better
to keep it as default in order to not confuse developers.

All uib's templates were moved to appropriate index files
because there is no other options to replace default path to
the uib's template folder.

Change-Id: I858fadb93b0e3f2ecac9cff15fae76be8c514c29
Reviewed-on: http://review.couchbase.org/68162
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>