MB-21281: Add notion of "UI compat version"
This change adds the idea of a "UI compat version" to the UI
server-side infrastructure. The UI compat version works in the
following way:
(1) A new version will be introduced when the UI changes
    significantly enough that we don't believe existing UI components
    will be able to work in the new UI without change
    (i.e. an incompatible change)
(2) The version will be set equal to the compat version of
    the release that the UI first goes out in. So, the Spock UI will
    have a UI compat version of [5, 0]. The Watson UI compat version
    will be [4, 5]. What we currently call the "classic UI", the
    UI in 4.1 and earlier, will have a UI compat version of [4, 1].
    In a sense the old "cells-based" UI should probably have a compat
    version of [3, 0] or perhaps even [2, 5], but it's a moot point
    what we call it since it's been superseded by the watson UI
    and will be by the spock UI.

Also this change refactors the 3 different functions written to
return the UI root for different UI compat versions into a single
function.

Change-Id: Ib61d9c2d5d30f3a466862314342e9618f863335e
Reviewed-on: http://review.couchbase.org/68552
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>