MB-21281: Add pluggable UI support for UI-compat-versions
With this change pluggable UI components can tell
ns_server where to find the head.frag.html file for different
UI compat versions. This allows pluggable UI components to use
different web resources for different UI version and to adapt
to whatever directory structure is convenient.

A new version-dirs attribute is added to the pluggable UI
config. E.g.
  "version-dirs": [
      {"version": [4, 5], "dir": "."},
      {"version": [5, 0], "dir": "spock"}
  ],

This tells ns_server that the head.frag.html for 4.5 is
to be found under the $doc-root/. (where $doc-root is the
component's specified doc root) and under
$doc-root/spock for 5.0. The pluggable component writer is
free to choose directories that work well for them and
to create head fragment files that share resources.

If no version-dirs attribute is explicit provided,
{"version": [4, 5], "dir": "."} is assumed by default (this
provides backward compatibility for existing pluggable UI
configs.)

Change-Id: Ia9947872385a62980c0f81f83b8d34beebff3a22
Reviewed-on: http://review.couchbase.org/68553
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>