Make sure that stop_ns_server works if server is already stopped.
This is specifically needed for the cluster leave logic to work in the
following case. If leave fails somewhere after it stopped the
ns_server, ns_cluster will try again after it restarts. But now the
attempt to shutdown the server will fail. It's probably the case that
eventually entire ns_server_cluster_sup will be stopped because of max
restart intensity limit being reached. Only after that will ns_cluster
be able to complete the leave (because entire ns_server vm will be
restarted, and ns_server_nodes_sup will be started again too).

Change-Id: I3109a7a47802bbd648ad5d8ec75e49fe0351d1a4
Reviewed-on: http://review.couchbase.org/68977
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>