move base into app
I believe we have to get rid of that folder since some Spock changes
in "base" folder require add changes in Watson in order to not break
the last one. Also, as far as I know we are not going to add new features
in Watson. So, let's keep them completely separately.

Change-Id: Id9e63dab11866801d19ba7b544ea50d35f2600b7
Reviewed-on: http://review.couchbase.org/69751
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>