Remove unnecessary sync_announcements from delete_bucket.
Git history suggests that it was there as an early attempt to make
sure that bucket deletion is synchronous. But it actually doesn't
guarantee much: the events are processed asynchronously. In addition,
we do have a logic that waits until bucket is gone on all nodes. So
there's no need for this sync_announcements call.

Change-Id: Idde7b5e203146af56fe73d4fa22b0f58fadcd36f
Reviewed-on: http://review.couchbase.org/70221
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>