[GNUmakefile] Pretend couchdb projects live in our deps dir.
In addition streamline their layout to match what typically is used in
Erlang projects: toplevel directory for different meta stuff and src,
include and ebin subdirectories.

All of this is only helpful for those who use edts for code
navigation. It only recognizes properly structured projects. So the
"streamlining" allows it to find couchdb sources and binaries. Of
course, I could actually fix couchdb layout, but that would only solve
the problem for spock and future releases. Whereas this hack I can
easily use with older releases too.  In addition, I didn't want to touch
to much of a cmake code.

Change-Id: Id1d277c3b3a3045dc8d107085e9c3e75594dadea
Reviewed-on: http://review.couchbase.org/70134
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>