Don't pull config from remote node before merging dynamic config.
It was there to prevent conflicts on {node, node(), membership} that
is set with vclock as part of merging in of dynamic config. Instead we
just set it using set_initial option, like we do with some other keys.

Generally, the set of per-node keys that can be modified remotely is
limited, and I believe it's better to be explicit about them.

Change-Id: I5ec48c8c4acd6a33e516a7187bbba53105622d8e
Reviewed-on: http://review.couchbase.org/70211
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>