Add function to compute same URL on different node ...
... running a specified service.

This is needed for cases where a UI page can't be presented because
a required service is running on a different node. Query, FTS and
views all have need of this functionality and all do it differently --
it's time we unified.

The function computes the URL for the same $location on a different
node. It retains the same logic around correctly handling
$locations under HTTPS as the old code.

Change-Id: I81f9c6691bf6efd9157cf753febadeb727a7e219
Reviewed-on: http://review.couchbase.org/71610
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>