[cmake] Allow creating test targets that must be run explicitly.
That is, they don't automatically become part of the tests that get
run if you do "make test" in either top-level or ns_server/build
directories.

To showcase the use, I added a target to execute just the eunit
tests (which is, as of now, not different from executing all tests).

Change-Id: I1006e2617a310ad2579e080c84c5add392c3ff6a
Reviewed-on: http://review.couchbase.org/71700
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>