MB-20206: Fix memcached log order on Windows
Adds the `/tw` flag to the `dir` command to ensure that the logs
are sorted with the least recently written file first. This ensures a
match with the UnixTask which uses `-tr`. As far as I can tell the
default is to order based on the creation time which is not always what
we want.

Change-Id: Ia6551326ef8e0d0f2dba6baac075ad45158aac79
Reviewed-on: http://review.couchbase.org/73806
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Oliver Downard <oliver.downard@couchbase.com>