MB-22208: Reprovision buckets upon memcached restart.
The following are the changes:

[1] The orchestrator has been changed to pick up all the
janitor items first and then call ns_janitor:cleanup on them
one at a time. Earlier we should to spawn a sub-process for
every janitor item but now we just use on sub-process to
cleanup all the janitor items.

[2] The orchestrator gets a list of unsafe nodes, if any,
from the janitor which it then passes to the auto reprovision
module to promote the respective REPLICAs for affected
vbuckets.

[3] The janitor cleanup now looks for unsafe nodes. These
are the nodes on which the memcached process has restarted
within auto-failover timeout and by virtue of which the
ACTIVEs are found to be in missing state for certain vbuckets.
If unsafe nodes are found then the janitor doesn't bring
the bucket online at all but simply return the list of
unsafe nodes back to the orchestrator.

[4] The auto-reprovision module now hosts a new API called
"reprovision_buckets" which takes in a list of ephemeral
buckets and unsafe nodes. It then promotes the REPLICAs
and persists this information in the bucket config. The
bucket will be brought online during the next janitor run.
This API also updates the auto-reprovision count and if the
count exceeds the max_nodes configured then disables the
feature until the count is reset. The actual ns_config will
be updated with latest bucket configs and auto-reprovision
count as a single transaction.

Change-Id: Id837767d627b5610739cd5244d37b6c007961ae7
Reviewed-on: http://review.couchbase.org/76398
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>