MB-22208: Add auto-reprovision handling module.
PROBLEM STATEMENT: With ephemeral buckets around, if the
memcached process restarts on any node within the auto
failover timeout, then we will bring the bucket online on
that node which can lead to data-loss as the ACTIVE copy will
have no data upon restart and the REPLICA will also lose
data when the DCP connections are re-established.

SOLUTION: To handle this case, the janitor during cleanup
will look for the condition where a vbucket's MASTER is
marked as missing. These nodes are dubbed as 'unsafe' nodes.
If the janitor finds any such instance then the bucket will
not be brought online immediately and it'll will pass a list
of all unsafe nodes found back to the orchestrator. The
orchestrator will use this information to reprovision the
bucket by promoting a REPLICA of the respective vbuckets and
then bring the bucket online.

This change introduces the following:
[1] A new module which would be responsible for automatically
    reprovisioning the buckets given a list of unsafe nodes.
    This change only introduces the configuration related APIs.

[2] We store the configuration info in ns_config and a default
    config has been added to ns_config_default module. The auto
    reprovision config contains the following params:
    [a] enabled - indicates whether the feature is enabled or not.
    [b] max_nodes - indicates max # of auto-reprovision operations
                    that can be performed.
    [c] count - the current number of auto-reprovisions performed.

Change-Id: I6f7cd33506f6696870db9770a14bf7ec446b32d9
Reviewed-on: http://review.couchbase.org/76396
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>