Always cancel stop timer on rebalance completion.
There's no guarantee that rebalancer actually terminates with reason
'stopped' when we ask it to stop: it might crash, or we might have
asked that just around the time when it successfully completed. That
would result in ns_orchestrator receiving spurious timeout
messages. It's even possible that it would terminate a subsequent
rebalance.

Change-Id: Idfd111f244e195bf8ee2250ba9b11e934b833b47
Reviewed-on: http://review.couchbase.org/73534
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>