[cmake] Make ns_couchdb depend on ns_server.
Any of our erlang projects can be built (but not used) independently
of each other. For that reason ns_server does not have to depend on
the subprojects. But since ns_server also serves as a place for common
code (and this common code is occasionally is a behavior), it makes
sense to build ns_server before certain subprojects (those who want to
use the behaviors defined in ns_server).

Change-Id: Ib420e4961af066cc737dddc6fc39ac93c4eeb8ae
Reviewed-on: http://review.couchbase.org/73602
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>