Expect that we still don't have a connection in unset_rebalance.
It might happen in the following scenario. A one node cluster is
initialized with topology aware service, orchestrator will try to run
initial rebalance (as part of service janitoring) for the
service. That happens shortly after starting service_agent and the
corresponding service itself. If while agent waits for the connection
new node is added and rebalanced in, then orchestrator will terminate
the janitor run to proceed with rebalance. That will result in
unset_rebalance call when connection is still missing. In this case we
don't have long poll workers yet, so we shouldn't attempt to
terminate/restart them.

Change-Id: I4f5b0a691755557eda1729de88a6f67361788395
Reviewed-on: http://review.couchbase.org/73683
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>