MB-22814 Sync config to couchdb when starting bucket supervisor.
Since config replication (even to local nodes) is asynchronous, it's
possible that when we try to start processes on couchdb node, it
hasn't seen the config for the bucket yet. Depending on a particular
process, it might or might not result in failure. So we explicitly
synchronize config to couchdb node here.

Change-Id: I20735752c581b12be231fbe8c3bec2e6a55ef738
Reviewed-on: http://review.couchbase.org/73728
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>