MB-19339: Add support for privilege_debug for memcached
Memcached supports privilege_debug mode to have the system
approve a request when the RBAC system would deny access.
The motivation for this mode is to be able to run an
application so that one may figure out which privileges
it need.

This mode should _NEVER_ be enabled in production!

To enable privilege debug mode the following command
may be used:

curl -u Administrator:asdfasdf \
           -X POST \
           http://127.0.0.1:8091/pools/default/settings/memcached/node/self \
           --data privilege_debug=true

Change-Id: Ia40d7809e422c709421b37b0d7980164b8edae76
Reviewed-on: http://review.couchbase.org/73949
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>