Invalidate dead lock holders.
We don't expect that to happen, but it's possible if a process get
killed in an unfortunate moment.

Otherwise, if such thing happens, the tables end up in unknown but
consistent state.

A note on testing. It proved to be harder than I expected to write an
automated test for this without deliberately introducing delays in the
code. For that reason there are no automated tests as of now, but I
did test it locally with those delays. I might revisit this later.

Change-Id: I66633fae57717707b6646cc6e755b975af0d36c0
Reviewed-on: http://review.couchbase.org/74364
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>