Enable XATTRs in spock clusters
The idea is that XATTRs are enabled by default in spock clusters, just
like subdoc or other memcached commands that get introduced from time to
time.

The patch is simple: if the cluster is at compat version spock we enable
XATTRs in the memcached config and add it to the bucket capabilities.
XATTRs are also supported on ephemeral buckets as I don't see why they
shouldn't be.

Merging this patch will cause MB-22163 to get reopened and a discussion
will commence on how we can mitigate the impact to write performance
in all use cases. This discussion may lead us to a place where XATTRs
are not enabled by default on all buckets and thus a change to this
patch. However, the best thing at this time is to merge the patch,
reopen the performance ticket and debate how we're going to tackle
it.

Change-Id: I5be8916dfae16c1c9d9781e8bea2499792d089bd
Reviewed-on: http://review.couchbase.org/74214
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>