MB-22909: Support XATTR replication.
In order to support replication of XATTRs we need to negotiate
the XATTR feature on the target node when we are setting up the
DCP proxy. If the XATTR feature is enabled on the target node
then we need to set the XATTR bit while issuing an open conn
DCP command.

During upgrade the replications will be setup without XATTRs
and when the cluster becomes Spock then we need to reset the
connections in order to enable XATTRs on all the replications.

In the mixed mode, the expectation is that a client is not
allowed to set XATTRs and any attempt to do so will result in
a failure.

Change-Id: Ibd9aacedd6028516eafa5d76edf5a26bd635a013
Reviewed-on: http://review.couchbase.org/74352
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>