MB-22940: Accept purgeInterval param while...
...creating/updating ephemeral buckets.

For couchbase buckets the purge interval is accepted as part of
the overidden auto compaction settings. But for ephemeral buckets
overiding auto compaction settings doesn't make sense. So the
approach taken is to parse the purge interval value separately
from the auto-compaction settings. For couchbase buckets it'll be
accpeted only if auto-compaction setting is overidden and for
ephemeral buckets it'll be accepted only if the auto-compaction
setting is not overidden.

Change-Id: I5cc15a8969cd2047b309a994f0da9fcb29ad1cca
Reviewed-on: http://review.couchbase.org/74813
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>