MB-23260 put bucket sasl_password's related info into
...isasl.pw and memcached.rbac under the "<bucket name>;legacy" key

The "bucket users" will not share the namespace with the normal users,
and in order for the old clients which used the bucket users
(XDCR for instance) we need to make a way for that to work.
The current design is that ns_server provides these to memcached
as "bucketname;legacy" with the bucket password.
Memcached should (for PLAIN auth) first check that password if
it exists, and if it doesn't it should fall back to the other user.

corresponding memcached change:
http://review.couchbase.org/74966/

Change-Id: I965c84cd076200f4441718bfee2e5399a2e92c0e
Reviewed-on: http://review.couchbase.org/75029
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>