MB-22208: Add gen_server to handle janitor state.
The attempt here is to move the janitor cleanup handling done in
the orchestrator into a new module thereby making orchestrator
module simpler.

The cleanup process is still completely driven by the orchestrator.
Just the state maintainence is moved to a separate module.

NOTE: There is one place-holder API implemented in the orchestrator
      to keep the dialyzer happy. This will be filled up as part of
      the  orchestrator and janitor_server intergration.

Change-Id: Ibe57e38a8238607a8305446b61ce4efeeae1b8bd
Reviewed-on: http://review.couchbase.org/76521
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>