MB-22208: Add janitor server to supervision...
..tree.

The ns_janitor_server and ns_orchestrator modules have been
grouped under a new supervisor called ns_orchestrator_child_sup.
This is because these 2 modules are interdependent and when one
dies the other one should not be around.

Also, the auto-failover module needs to be restarted when
ns_orchestrator restarts. This is because the auto-failover
module calls ns_orchestrator module's 'try_autofailover' API
which waits if the ns_orchestrator module is not around. This
can lead to multiple 'tick' timer messages sitting in the
auto-failover module's mailbox which can result in failing
over a node prematurely. This problem is altogether handled
by setting the 'intensity' and 'period' parameters of
the ns_orchestrator_child_sup module to '0' and '1'
respectively. This ensures that all the modules under the
ns_orchestrator_sup module restarts when any process under the
ns_orchestrator_child_sup module restarts.

Change-Id: Id817682b204b7fd95333d2407ed7f72a3fd3df5d
Reviewed-on: http://review.couchbase.org/76522
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>