MB-22208: Clear reprovision count after rebalance.
We allow auto-reprovisioning to be done on 'max_nodes' #
of nodes & we track the actual number of nodes that have
been auto-reprovisioned in ns_config. If we have reached
the 'max_nodes' limit then we don't perform any more
auto-reprovisioning. This count needs to be cleared after
the buckets are rebalanced.

Change-Id: I3097b8d09f3ebacd818fb82aaa65f7e755cb6be4
Reviewed-on: http://review.couchbase.org/76770
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>