MB-23748: Allow index mode to be altered when...
...there no active index nodes.

Currently we just check for a list of nodes running the index service
and if it's non-empty we don't allow the index storage mode to be
changed. But the problem with this is even the node in 'inactiveAdded'
state will be accounted for and setting of storage mode fails. This is
especially a problem when the first index node is being added to the
cluster and the storage mode is not initially set.

The fix is to look for index nodes (currently running + failed over
nodes and down nodes) and if there are none then allow the storage
mode to be changed.

Change-Id: I670f58a3a29563f595db4131ed3a516d64050a59
Reviewed-on: http://review.couchbase.org/77386
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>