Catch all in ns_memcached:warmed() and return false
The semantics of warmed() used to be that if a timeout occured
it returned false, that is, the bucket was considered
not ready. This was changed to not catch the exception in
0f1f744c8cdd7a3b8403adcbe79bca6edd25939a /
http://review.couchbase.org/#/c/73598. A result of this change
was that if the timeout occurs, ns_heart crashes, which probably
wasn't intended and isn't desirable.

This patch catches exceptions in warmed() and returns false.

This patch also changes kv_monitor to not catch timeout exits
in the call to warmed_buckets(), since they are no longer
thrown. An additional affect is that if timeouts do occur
only the calls to buckets that timeout will be considered
not ready instead of all buckets being considered not ready.

Change-Id: I5ec858ae88299bf89b5ccd86bc7f8306aeef0a0a
Reviewed-on: http://review.couchbase.org/78221
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>