MB-22627: Add XDCR to bucket total ops definition.
The total ops on the bucket contain ops contributed by XDCR metrics but the
hover on definition in the UI currently does not reflect this. To fix this we
add the XDCR metrics which are ep_num_ops_del_meta, ep_num_ops_get_meta and
ep_num_ops_set_meta and also explicitly mention XDCR in the definition so that
users are aware that it is included in the total ops.

Change-Id: Ic530b2ab9a273b96958aecfb0d2d483a53b57b0f
Reviewed-on: http://review.couchbase.org/78448
Tested-by: Amit Krishnan <krish.amit@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>