[meck] Pull meck source into ns_server
Meck is an Apache 2.0 licensed mocking library for Erlang.
It looks to have reasonable features and to be the best of the
currently available open source Erlang mocking libraries.

Using meck will allow us to get rid of `mock` and `mock_gen_server`.
Meck includes all the module mocking capabilities of mock but
importantly also includes the ability to mock some functions but
"passthrough" to the original module for unmocked functions.

Mock_gen_server is no longer needed - it's easy to use meck to
create a module for scratch, add an init method and start a
gen_server in a couple of lines of code. Overall this means
we have one mocking API that's necessarily consistent.

A gentle but informative introduction to meck can be found here:
http://www.erlang-factory.com/upload/presentations/422/MeckatErlangFactory,London2011.pdf

The v0.8 release notes are also worth reading:
https://github.com/eproxus/meck/wiki/0.8-Release-Notes

And of course, full details are in the source code at:
https://github.com/eproxus/meck

Change-Id: Ie14c57cd981418f043a7b6de750d6b2b73e96373
Source: https://github.com/eproxus/meck
Reviewed-on: http://review.couchbase.org/78677
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>