MB-24420: Perform check_config in a separate process.
The check_config ends up calling the memcached stats API
which occassionally takes a very long time to complete.
This will result in blocking up the ns_memcached process
which can ultimately lead to failover. The fix is to perform
the check_config operation in a separate process using a
socket from the ns_memcached_sockets_pool.

Change-Id: I2026e6dd3066656338565da43dc1c0b69f352645
Reviewed-on: http://review.couchbase.org/78944
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>