MB-24648: Fail node add when its already part of cluster.
Currently if a node is present in the nodes_wanted list and for some
reason its config gets deleted we allow the node to be re-added to
the cluster without doing a failover first. This causes data loss.
With this changeset we check first if the node is present in
nodes_wanted and add only if it doesn't already exist.

Change-Id: Id3367e1a0865db245e695d0e5134904432cc642a
Reviewed-on: http://review.couchbase.org/79972
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Amit Krishnan <krish.amit@gmail.com>
Reviewed-by: Poonam D <poonam@couchbase.com>