MB-24751: Consider active ephemeral buckets while...
...deleting unused db files.

ns_server tries to delete the unused (deleted) DB files by getting
a list of active buckets from the config and comparing that against
the list of bucktes found on disk. The buckets that are present on
disk but that are not active are deleted.

The problem here is that ephemeral buckets are not picked up as part
of the active bucket list and this makes DB files deletion logic to
think that ephemeral buckets are unused even though they are active.

The fix is to pick up the ephemeral buckets as part of the active
list.

Change-Id: I682cd9d9ed7290a173d0976d7150abc68222f88b
Reviewed-on: http://review.couchbase.org/81186
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Poonam D <poonam@couchbase.com>