MB-25372: Don't consider couchbase buckets...
...while determining the unsafe nodes.

The janitor, while performing the cleanup operation, checks to see
if any node is unsafe (node on which the active vbuckets are found
to be missing). This is required to trigger the auto-reprovision
operation for ephemeral buckets. But currently this is being done
even for couchbase buckets. So when, for some reason, the data files
are missing for a couchbase bucket the janitor triggers the auto
reprovision operation which can lead to an infinite loop.

This patch fixes the problem by making the janitor look for unsafe
nodes only if the bucket is ephemeral. If the bucket is not
ephemeral then the bucket will be brought online.

Change-Id: If8f999582e61d2bebd015e98064fb728f899798b
Reviewed-on: http://review.couchbase.org/81353
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>