MB-23675: setupServices API to take an extra...
...parameter called 'setDefaultMemQuotas'.

In the cluster setup wizard, if the user chooses to configure the
memory quotas and enters values that are greater than the maximum
usable memory, then the wizard shows an appropriate error message
and stops the user from completing the setup. At this point, if
the user goes back to choose "Finish With Defaults" option that
ends up failing too.

This is because the user provided memory quotas in the manual
config step gets stored in the cluster config and the "setupServices"
API called during the "Finish With Defaults" step tries to use
quota information present in the cluster config and fails.

The fix is to add new parameter to "setupServices" API using which
the user can indicate to recompute the defaults and use them instead.

Change-Id: I1eae911e46cbcb967242314ededc03c1a7e6e191
Reviewed-on: http://review.couchbase.org/81867
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>