MB-25659: Add a timeout for cbcollect_info 'ss' stats command
Add a timeout of 300 seconds to Linux Task 
'ss -an --info --processes'. This is due to the fact we have noticed
that this task can occasionally hang and lead to log collection taking
hours in some instances. It is much better to bail out of the task if
it is taking a long time, instead of not getting logs due to the process
being killed by a user thinking it is inactive.

Change-Id: Ifacb26cbc4add26587cea2370b1f105f13782636
Reviewed-on: http://review.couchbase.org/82203
Tested-by: Tim Bradgate <tim.bradgate@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>