MB-25663: Add a log call for cbcollect_info timed out commands
Add a log call to the execute function in the Task class to 
add information into cbcollect_info.log to warn that the command was 
forcibly timed out. This is useful when running the cbcollect_info.log
file through our interal Support tools, as it makes it easy to see which
(if any) commands were foribly timed out due to taking an excessive time
period to complete. This can further be used to optimise cbcollect_info
in the future, by looking at patterns of commands being timed out and
potentially removing them in future releases if not deemed useful.

Change-Id: Iefd0eaa4dfb38b12effd1c32a0e3a8ba873176da
Reviewed-on: http://review.couchbase.org/82213
Tested-by: Tim Bradgate <tim.bradgate@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>