MB-10409: Handle URL construction.
For IPv6 configured hosts the respective URLs need to be constructed
with the IP address enclosed within '[' and ']' brackets.

In ns_server, we construct a lot of URLs using the localhost address.
This change makes use of the abstraction APIs defined to construct
the URLs appropriately.

Change-Id: I2af21b4908aa54f8feb98569eb9ebe237fec484f
Reviewed-on: http://review.couchbase.org/71871
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>