[cut] Add cut parse transform.
This is part of erlando[1] project and it implements poor man's
partial application for Erlang. Please see the readme file at the
project's page for the details on how to use it.

To play it safe, the parse transform is opt-in, that is, you need to
enable the parse transform on a per-file basis. You can do this by
simply including the cut.hrl header file.

[1] https://github.com/rabbitmq/erlando

Change-Id: I9fa04482a2e29c9815cc5f7312116ae223eb33f5
Reviewed-on: http://review.couchbase.org/84706
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>