MB-10409: Alter 'is_good_address' API to handle...
...an IPv6 address as a parameter.

The current implementation looks for the presence of "." to
determine if 'Address' passed in is a hostname. If an IPv6 address
if passed to this function it would fail with a wrong message. The
fix is to take IPv6 address into account and error out appropriately.

Also, if the Couchbase server is configured to run in IPv6 mode
and the user passes in an IPv4 address an approrpiate error message
will be returned.

Change-Id: I4914e0a1ae857676e2d227c3304097d011fbd336
Reviewed-on: http://review.couchbase.org/83748
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>