MB-10409: Fix 'call_port_please' API to consider...
...IPv6 protocol when fetching port info from EPMD process.

When the 'Host' parameter is an FQDN (or hostname) the epmd:port_please
API implementation uses "inet" protocol by default. This will fail
if the host is configured with IPv6. But if we pass in the IP Address
instead of hostname the EPMD API does the right thing.

So we first get the address of the 'Host' based on the network
family configured and pass the returned IP address as a parameter
to the EPMD API.

Change-Id: Iaaeab868e8c910d0c1fc991f13a73d76e064478b
Reviewed-on: http://review.couchbase.org/83749
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>