MB-26206: cbcollect_info: fix recursive loop in /sys/block
When dumping the block device settings in cbcollect_info; if there's a
symlink in one of the directories that points back to the same
directory (or higher up), the 'grep -R' will loop indefinitely.

Fix by changing to '-type f' in find. Also flip to xargs for
multi-file processing as it's faster than '-exec' (0.01s vs 0.3s on
mancouch).

Change-Id: I12f212fa954e7b033574a4eb708ffa1128af14d0
Reviewed-on: http://review.couchbase.org/83869
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>