MB-26257: cluster_run to allow node renaming...
...feature when used in IPv6 mode.

The cluster_run script uses the Google Public DNS to
identify the IP address of the machine on which cluster_run
is being run and employs the resulting address to rename
the node. This was disabled for IPv6 mode as raw IPv6
addresses were not supported earlier.

This fix uses the Google Public DNS's IPv6 endpoint to
determine the IPv6 address of the machine.

Change-Id: If5db84c5b857ae6984f9ee8a3c5a18f0af173796
Reviewed-on: http://review.couchbase.org/84583
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>