MB-26956: Make 'path' field compulsory when...
...client-cert-auth 'state' is being set to 'enable' or 'mandatory'.

When the user sets the 'state' of client certificate authentication
settings to either 'enable' or 'mandatory', then we would also want
the user to specify the 'path' setting. Because the 'path' setting
determines the certificate field from which the server will have to
extract the user name from. Without the path setting, the server
will not be able to authorize the user.

Change-Id: I687b7578a61feac6e84ac89ffce5f58f22cb5941
Reviewed-on: http://review.couchbase.org/86251
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>