[mb_master] Get rid of the worker state.
It's not a normal situation that there's an mb_master process running
on a node while it's not in nodes_wanted. That can only happen when
the node is being removed from the cluster. So for the simplicity sake
we'll just run as a candidate in such case.

Change-Id: I2ea547e35510c2eb3df2ca5532d20c1f7d204f83
Reviewed-on: http://review.couchbase.org/86327
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>