MB-26479 This reverts commit 5bce08e3c63544d8f32397ab50564ac886d448d2, and reintroduces commit 6b190bad557ce6a06520041ddb606610c1400c50, after enabling corresponding view engine changes.
Pick the CAPI bind_address based on the networking mode
in which the cluster has been configured. With the support for IPv6
introduced we are now adding separate bind address parameters in capi.ini file.
Depending on the networking mode in which the cluster has been configured,
corresponding bind address will be chosen.

Change-Id: Ie6090139bbe24bfcc6ece8f5d54d7d90dfd0413c
Reviewed-on: http://review.couchbase.org/86903
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>
Tested-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>