MB-27111: Change is_good_address API to check...
..if the supplied FQDN/Address is IPv6 capable.

Currently ns_server enforces that the given Address/FQDN doesn't
resolve to an IPv4 address when the cluster is configured in IPv6
mode. This can lead to problems on dual stack hosts with FQDNs
mapped to both IPv4 & IPv6 addresses.

The fix here is to remove the enforcement and just make sure that
the given FQDN/Address is IPv6 capable. This simple check should
be enough as all the components within ns_server will be configured
to use either IPv4 or IPv6 protocol.

Change-Id: Id6bd4fdbd51191e92e3c2bf2d5da01963d1e1a6b
Reviewed-on: http://review.couchbase.org/86800
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>